update serial absl status

+++ ros2/encoder_publisher.cc
@@ -28,7 +28,7 @@ public:
     
     for (const auto& single_perception : config.robot().perceptions().single_perceptions()) {
       if (single_perception.perception_type() == robot::perception::PerceptionType::ENCODER && 
-          single_perception.node_id() == node_id) {
+          static_cast<int>(single_perception.node_id()) == node_id) {
         const auto& encoder_proto = single_perception.encoder();
         
         // First, create the interface and check if it's valid.
